+++ src/eca/llm_api.clj
@@ -8,6 +8,7 @@
    [eca.llm-providers.ollama :as llm-providers.ollama]
    [eca.llm-providers.openai :as llm-providers.openai]
    [eca.llm-providers.openai-chat :as llm-providers.openai-chat]
+   [eca.llm-providers.openrouter]
    [eca.llm-util :as llm-util]
    [eca.logger :as logger]))