+++ src/eca/llm_api.clj
@@ -9,6 +9,7 @@
    [eca.llm-providers.openai :as llm-providers.openai]
    [eca.llm-providers.openai-chat :as llm-providers.openai-chat]
    [eca.llm-providers.openrouter]
+   [eca.llm-providers.deepseek]
    [eca.llm-util :as llm-util]
    [eca.logger :as logger]))