+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Support more providers login via `/login`.
+  - openai
+
 ## 0.46.0
 
 - Remove the need to pass `requestId` on prompt messages.